Refactored VTG parser

+++ src/parsers/gpvtg.h
@@ -3,14 +3,14 @@
 
 #include <stdlib.h>
 #include <string.h>
-#include <time.h>
 #include <nmea.h>
 
-typedef struct {
+typedef struct
+{
 	nmea_s base;
-	double track_deg;
-	double gndspd_knots;
-	double gndspd_kmph;
+	float track_deg;
+	float gndspd_knots;
+	float gndspd_kmph;
 } nmea_gpvtg_s;
 
 /* Value indexes */
@@ -18,4 +18,4 @@ typedef struct {
 #define NMEA_GPVTG_GNDSPD_KNOTS 4
 #define NMEA_GPVTG_GNDSPD_KMPH 6
 
-#endif  /* INC_NMEA_GPVTG_H */
+#endif /* INC_NMEA_GPVTG_H */
